pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monkey_do"
version = "0.1.0"
description = "A full-stack data management and processing system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.164.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "gspread>=6.2.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools]
packages = ["monkey_do"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
